requirements.txt
requests==2.32.3
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.3
python-dotenv==1.0.1
nba_api==1.5.1
pymc==5.16.1
arviz==0.18.0
numpy==2.0.1
